+++ src/converter.ts
@@ -16,14 +16,6 @@ export interface StateConverterOptionsWithContext
 export interface ConverterOptions<R, V> {
   convert(raw: R, options: StateConverterOptionsWithContext): V;
   render(value: V, options: StateConverterOptionsWithContext): R;
-  rawValidate?(
-    value: R,
-    options: StateConverterOptionsWithContext
-  ): boolean | Promise<boolean>;
-  validate?(
-    value: V,
-    options: StateConverterOptionsWithContext
-  ): boolean | Promise<boolean>;
   emptyRaw: R;
   emptyValue?: V;
   emptyImpossible?: boolean;
@@ -39,22 +31,20 @@ export interface IConverter<R, V> {
   convert(
     raw: R,
     options: StateConverterOptionsWithContext
-  ): Promise<ConversionResponse<V>>;
+  ): ConversionResponse<V>;
   render(value: V, options: StateConverterOptionsWithContext): R;
   defaultControlled: Controlled;
   neverRequired: boolean;
   preprocessRaw(raw: R, options: StateConverterOptionsWithContext): R;
 }
 
-export class ConvertError {}
-
 export class ConversionValue<V> {
   constructor(public value: V) {}
 }
 
-export type ConversionError = "ConversionError";
-
-export const CONVERSION_ERROR: ConversionError = "ConversionError";
+export class ConversionError {
+  constructor(public type: string = "default") {}
+}
 
 export type ConversionResponse<V> = ConversionError | ConversionValue<V>;
 
@@ -92,34 +82,16 @@ export class Converter<R, V> implements IConverter<R, V> {
     return this.definition.preprocessRaw(raw, options);
   }
 
-  async convert(
+  convert(
     raw: R,
     options: StateConverterOptionsWithContext
-  ): Promise<ConversionResponse<V>> {
-    if (this.definition.rawValidate) {
-      const rawValidationSuccess = await this.definition.rawValidate(
-        raw,
-        options
-      );
-      if (!rawValidationSuccess) {
-        return CONVERSION_ERROR;
-      }
-    }
+  ): ConversionResponse<V> {
     try {
       const value = this.definition.convert(raw, options);
-      if (this.definition.validate) {
-        const rawValidationSuccess = await this.definition.validate(
-          value,
-          options
-        );
-        if (!rawValidationSuccess) {
-          return CONVERSION_ERROR;
-        }
-      }
       return new ConversionValue<V>(value);
     } catch (e) {
-      if (e instanceof ConvertError) {
-        return CONVERSION_ERROR;
+      if (e instanceof ConversionError) {
+        return e;
       }
       throw e;
     }